PublishService/requirements.txt
mongoengine==0.15.0
dnspython==1.15.0
boto3==1.4.6
awscli==1.11.133
watchtower==0.4.0
requests==2.18.4
python-dotenv==0.6.5